[NOID] added gradle deps
