+++ full/build.gradle
@@ -106,6 +106,7 @@ dependencies {
     }
 
     def withoutJacksons = {
+        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-core'
         exclude group: 'com.fasterxml.jackson.core', module: 'jackson-annotations'
         exclude group: 'com.fasterxml.jackson.core', module: 'jackson-databind'
     }
@@ -139,6 +140,8 @@ dependencies {
 
     implementation group: 'com.opencsv', name: 'opencsv', version: '5.7.1'
     implementation group: 'commons-beanutils', name: 'commons-beanutils', version: '1.9.4'
+    implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.17.0', withoutJacksons
+    
     compileOnly group: 'org.ow2.asm', name: 'asm', version: '5.0.2'
 
     // schemacrawler
